backend/base/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Transaction


class TransactionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Transaction
        fields = '__all__'
