pyproject.toml
[tool.poetry]
name = "denfinancetracker"
version = "0.1.0"
description = ""
authors = ["Lawrence <lawrencesalinas7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.7.2"
django = "4.2.5"
django-cors-headers = "4.2.0"
djangorestframework = "3.14.0"
pytz = "2023.3.post1"
sqlparse = "0.4.4"
typing-extensions = "4.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
